// src/generated/models/saved-report-configuration-model.ts
import type { FieldDefinition } from './field-definition-model';
import type { SavedReportDefinition } from './saved-report-definition-model';

export const SavedReportConfigurationAggregationKeyToLabel = {
  'Count': 'Count',
  'Sum': 'Sum',
  'Average': 'Average',
  'Minimum': 'Minimum',
  'Maximum': 'Maximum',
  'None': 'None'
} as const;
export type SavedReportConfigurationAggregationKey = keyof typeof SavedReportConfigurationAggregationKeyToLabel;

export const SavedReportConfigurationDateScopeKeyToLabel = {
  'Last30Days': 'Last30Days',
  'Last90Days': 'Last90Days',
  'ThisYear': 'ThisYear',
  'Custom': 'Custom'
} as const;
export type SavedReportConfigurationDateScopeKey = keyof typeof SavedReportConfigurationDateScopeKeyToLabel;

export const SavedReportConfigurationVisualizationKeyToLabel = {
  'Summary': 'Summary',
  'Table': 'Table',
  'Bar': 'Bar',
  'Line': 'Line',
  'Donut': 'Donut'
} as const;
export type SavedReportConfigurationVisualizationKey = keyof typeof SavedReportConfigurationVisualizationKeyToLabel;

export interface SavedReportConfiguration {
  /**
   * @displayName Saved Report Configuration ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Configuration Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  configurationName: string;
  /**
   * @displayName Aggregation
   * @validationRule Required for create/update operations
   */
  aggregationKey: SavedReportConfigurationAggregationKey;
  /**
   * @displayName Custom Field Definition
   */
  customFieldDefinition?: Pick<FieldDefinition, 'id' | 'label'>;
  /**
   * @displayName Date Scope
   * @validationRule Required for create/update operations
   */
  dateScopeKey: SavedReportConfigurationDateScopeKey;
  /**
   * @displayName Dimension Key
   * @validationRule Required for create/update operations
   */
  dimensionKey: string;
  /**
   * @displayName Filters JSON
   * @validationRule Required for create/update operations
   */
  filtersJSON: string;
  /**
   * @displayName Layout JSON
   * @validationRule Required for create/update operations
   */
  layoutJSON: string;
  /**
   * @displayName Measure Key
   * @validationRule Required for create/update operations
   */
  measureKey: string;
  /**
   * @displayName Saved Report Definition
   * @validationRule Required for create/update operations
   */
  savedReportDefinition: Pick<SavedReportDefinition, 'id' | 'name1'>;
  /**
   * @displayName Sorting JSON
   * @validationRule Required for create/update operations
   */
  sortingJSON: string;
  /**
   * @displayName Visualization
   * @validationRule Required for create/update operations
   */
  visualizationKey: SavedReportConfigurationVisualizationKey;
}

export const _SavedReportConfiguration = 'SavedReportConfiguration' as const;