// src/generated/models/saved-report-definition-model.ts
import type { Person } from './person-model';
import type { Workspace } from './workspace-model';

export const SavedReportDefinitionSharingScopeKeyToLabel = {
  'Private': 'Private',
  'Workspace': 'Workspace',
  'Service': 'Service'
} as const;
export type SavedReportDefinitionSharingScopeKey = keyof typeof SavedReportDefinitionSharingScopeKeyToLabel;

export const SavedReportDefinitionStatusKeyToLabel = {
  'Active': 'Active',
  'Archived': 'Archived'
} as const;
export type SavedReportDefinitionStatusKey = keyof typeof SavedReportDefinitionStatusKeyToLabel;

export interface SavedReportDefinition {
  /**
   * @displayName Saved Report Definition ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  name1: string;
  /**
   * @displayName Access Permission Key
   * @validationRule Required for create/update operations
   */
  accessPermissionKey: string;
  /**
   * @displayName Created At
   * @validationRule Required for create/update operations
   */
  createdAt: string;
  /**
   * @displayName Description
   * @validationRule Required for create/update operations
   */
  description: string;
  /**
   * @displayName Is Deleted
   * @validationRule Required for create/update operations
   */
  isDeleted: boolean;
  /**
   * @displayName Owner Person
   * @validationRule Required for create/update operations
   */
  ownerPerson: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName Service Code
   */
  serviceCode?: string;
  /**
   * @displayName Sharing Scope
   * @validationRule Required for create/update operations
   */
  sharingScopeKey: SavedReportDefinitionSharingScopeKey;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: SavedReportDefinitionStatusKey;
  /**
   * @displayName Updated At
   * @validationRule Required for create/update operations
   */
  updatedAt: string;
  /**
   * @displayName Workspace
   * @validationRule Required for create/update operations
   */
  workspace: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _SavedReportDefinition = 'SavedReportDefinition' as const;